+++ pmfp/utils/project_rc_utils.py
@@ -58,4 +58,3 @@ DEFAULTRCSCHEMA = {
 
 # def write_rc():
 #     pass
-